server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["Ramon Meza <ramon_meza@live.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-fasthtml = "^0.4.4"


[tool.poetry.group.dev.dependencies]
black = "^24.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
